Reconsolidated tree helpers
